requirements-docker.txt
# Docker-compatible requirements (Python 3.10)
# Simplified version without strict version pinning

# Core
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.27.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# ML Models
scikit-learn>=1.2.0
xgboost>=2.0.0
lightgbm>=4.0.0
catboost>=1.2.0
statsmodels>=0.14.0

# Tinkoff API - using t-tech-investments from T-Bank repository
--extra-index-url https://opensource.tbank.ru/api/v4/projects/238/packages/pypi/simple
t-tech-investments>=0.3.0

# Visualization
matplotlib>=3.7.0
plotly>=5.0.0
seaborn>=0.12.0

# Technical Analysis
ta>=0.10.0

# Scheduling
APScheduler>=3.10.0

# Streamlit Dashboard
streamlit>=1.30.0

# HTTP client
requests>=2.28.0
aiohttp>=3.8.0

# Utilities
tqdm>=4.65.0
Jinja2>=3.1.0

grpcio>=1.70.0

